Move code from tokens view back to main
