+++ lib/views/tokens_view/tokens_view.dart
@@ -1,5 +1,4 @@
 import 'package:edumfa_authenticator/model/states/token_filter.dart';
-import 'package:edumfa_authenticator/utils/logger.dart';
 import 'package:edumfa_authenticator/utils/riverpod_providers.dart';
 import 'package:edumfa_authenticator/views/tokens_view/tokens_view_widgets/app_bar_item.dart';
 import 'package:edumfa_authenticator/views/tokens_view/tokens_view_widgets/connectivity_listener.dart';
@@ -8,14 +7,10 @@ import 'package:edumfa_authenticator/views/tokens_view/tokens_view_widgets/main_
 import 'package:edumfa_authenticator/views/tokens_view/tokens_view_widgets/tokens_list.dart';
 import 'package:edumfa_authenticator/views/tokens_view/tokens_view_widgets/tokens_list_filtered.dart';
 import 'package:edumfa_authenticator/views/view_interface.dart';
-
 import 'package:edumfa_authenticator/widgets/push_request_listener.dart';
 import 'package:edumfa_authenticator/widgets/status_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutterlifecyclehooks/flutterlifecyclehooks.dart';
-
-export 'package:edumfa_authenticator/views/main_view/main_view.dart';
 
 class TokensView extends ConsumerStatefulView {
   static const routeName = '/tokens';
@@ -29,21 +24,8 @@ class TokensView extends ConsumerStatefulView {
   ConsumerState<TokensView> createState() => _TokensViewState();
 }
 
-class _TokensViewState extends ConsumerState<TokensView> with LifecycleMixin {
+class _TokensViewState extends ConsumerState<TokensView> {
   final globalKey = GlobalKey<NestedScrollViewState>();
-
-  @override
-  void onAppResume() {
-    Logger.info('TokensView Resume', name: 'tokens_view.dart#onAppResume');
-    globalRef?.read(appStateProvider.notifier).state =
-        AppLifecycleState.resumed;
-  }
-
-  @override
-  void onAppPause() {
-    Logger.info('TokensView Pause', name: 'tokens_view.dart#onAppPause');
-    globalRef?.read(appStateProvider.notifier).state = AppLifecycleState.paused;
-  }
 
   @override
   Widget build(BuildContext context) {
